+++ packages/x-adapter-platform/src/__tests__/platform.adapter.spec.ts
@@ -124,9 +124,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/search?internal=true&query=chips&origin=popular_search%3Apredictive_layer&start=0&rows=0&sort=price+asc&filter=categoryIds%3Affc61e1e9__be257cb26&filter=gender%3Amen&filter=price%3A10.0-20.0&instance=empathy&env=test&lang=es&device=mobile&scope=mobile',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
     expect(response.totalResults).toBe(0)
@@ -195,9 +192,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/empathize?internal=true&start=0&rows=24&instance=empathy&env=test&lang=en&device=tablet&scope=tablet',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -246,9 +240,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/empathize?internal=true&query=boots&start=0&rows=24&instance=empathy&env=test&lang=en&device=tablet&scope=tablet',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -298,9 +289,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/empathize?internal=true&query=boots&start=0&rows=24&instance=empathy&env=test&lang=en&device=tablet&scope=tablet',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -351,9 +339,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/nextqueries/empathy?internal=true&query=makeup&scope=mobile&instance=empathy&device=mobile&env=staging&lang=en',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
     expect(response).toStrictEqual({
@@ -405,9 +390,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/relatedtags/empathy?internal=true&query=jeans&device=mobile&env=staging&lang=en&scope=mobile&instance=empathy',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
     expect(response).toStrictEqual({
@@ -445,9 +427,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/search/v1/query/empathy/skusearch?internal=true&query=jeans&origin=search_box%3Anone&start=0&rows=24&instance=empathy&env=staging&lang=en&device=mobile&scope=mobile',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -564,9 +543,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/topclicked?internal=true&start=0&rows=24&origin=search_box%3Anone&instance=empathy&env=test&lang=en&device=desktop&scope=desktop',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -598,9 +574,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/tagging/v1/track/empathy/click?filtered=false&follow=false&lang=en&origin=search_box%3Anone&page=1&position=1&productId=12345-U&q=12345&scope=desktop&spellcheck=false&title=Xoxo+Women+Maroon+Pure+Georgette+Solid+Ready-to-wear+Saree',
       {
         keepalive: true,
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
   })
@@ -637,9 +610,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/semantics-api/search_single/empathy?q=test&lang=en&instance=empathy&env=staging',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
 
@@ -681,9 +651,6 @@ describe('platformAdapter tests', () => {
       'https://api.staging.empathy.co/config/v1/public/configs?service=xcontrols&instance=empathy&env=staging',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )
     expect(response).toStrictEqual({
@@ -725,7 +692,6 @@ describe('platformAdapter tests', () => {
         method: 'POST',
         signal: expect.anything(),
         headers: {
-          'x-origin': expect.anything(),
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
@@ -776,9 +742,6 @@ describe('platformAdapter tests', () => {
       'https://search.internal.test.empathy.co/query/empathy/facets?query=books&origin=search_box%3Anone&filter=editorial%3AALFAGUARA&instance=empathy&env=test&lang=es&device=desktop&scope=desktop',
       {
         signal: expect.anything(),
-        headers: {
-          'x-origin': expect.anything(),
-        },
       },
     )