+++ packages/x-adapter-platform/src/endpoint-adapters/experience-controls.endpoint-adapter.ts
@@ -2,7 +2,7 @@ import type { ExperienceControlsRequest, ExperienceControlsResponse } from '@emp
 import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter'
 import { experienceControlsRequestMapper } from '../mappers/requests/experience-controls-request.mapper'
 import { experienceControlsResponseMapper } from '../mappers/responses/experience-controls-response.mapper'
-import { getConfigServiceUrl, getDefaultHeaders } from './utils'
+import { getConfigServiceUrl } from './utils'
 
 /**.
  * Default adapter for the experience controls endpoint.
@@ -18,9 +18,6 @@ export const experienceControlsEndpointAdapter = endpointAdapterFactory<
   responseMapper: experienceControlsResponseMapper,
   defaultRequestOptions: {
     id: 'experience-controls',
-    properties: {
-      headers: getDefaultHeaders(),
-    },
     parameters: {
       service: 'xcontrols',
     },