+++ packages/x-adapter-platform/src/endpoint-adapters/identifier-results.endpoint-adapter.ts
@@ -2,7 +2,7 @@ import type { IdentifierResultsRequest, IdentifierResultsResponse } from '@empat
 import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter'
 import { identifierResultsRequestMapper } from '../mappers/requests/identifier-results-request.mapper'
 import { identifierResultsResponseMapper } from '../mappers/responses/identifier-results-response.mapper'
-import { getDefaultHeaders, getSearchServiceUrl } from './utils'
+import { getSearchServiceUrl } from './utils'
 
 /**
  * Default adapter for the identifier results endpoint.
@@ -19,9 +19,6 @@ export const identifierResultsEndpointAdapter = endpointAdapterFactory<
   responseMapper: identifierResultsResponseMapper,
   defaultRequestOptions: {
     id: 'identifier-results',
-    properties: {
-      headers: getDefaultHeaders(),
-    },
     parameters: {
       internal: true,
     },