+++ packages/x-adapter-platform/src/endpoint-adapters/semantic-queries.endpoint-adapter.ts
@@ -2,7 +2,7 @@ import type { SemanticQueriesRequest, SemanticQueriesResponse } from '@empathyco
 import { endpointAdapterFactory, interpolate } from '@empathyco/x-adapter'
 import { semanticQueriesRequestMapper } from '../mappers/requests/semantic-queries-request.mapper'
 import { semanticQueriesResponseMapper } from '../mappers/responses/semantic-queries-response.mapper'
-import { getDefaultHeaders, getSemanticsServiceUrl } from './utils'
+import { getSemanticsServiceUrl } from './utils'
 
 /**
  * Default adapter for the semantic queries endpoint.
@@ -19,8 +19,5 @@ export const semanticQueriesEndpointAdapter = endpointAdapterFactory<
   responseMapper: semanticQueriesResponseMapper,
   defaultRequestOptions: {
     id: 'semantic-queries',
-    properties: {
-      headers: getDefaultHeaders(),
-    },
   },
 })