+++ packages/x-components/src/x-modules/ai/components/ai-carousel.vue
@@ -39,48 +39,84 @@
           </button>
         </span>
       </DisplayEmitter>
-      <slot name="sliding-panel" :suggestions="suggestionsSearch" :tagging="tagging">
-        <SlidingPanel
-          :class="slidingPanelClasses"
-          :scroll-container-class="slidingPanelContainerClasses"
-          :button-class="slidingPanelButtonsClasses"
-          :reset-on-content-change="false"
+
+      <AiGroupedCarousel v-if="group">
+        <template #sliding-panel-addons="{ arrivedState }">
+          <slot name="sliding-panels-addons" :arrived-state="arrivedState" />
+        </template>
+        <template #sliding-panel-left-button>
+          <slot name="sliding-panels-left-button" />
+        </template>
+        <template #sliding-panel-right-button>
+          <slot name="sliding-panels-right-button" />
+        </template>
+        <template #result="{ result }">
+          <slot name="result" :result="result" />
+        </template>
+      </AiGroupedCarousel>
+
+      <div v-else class="x-ai-carousel-suggestions" data-test="ai-carousel-suggestions-container">
+        <DisplayEmitter
+          v-for="{ query: suggestionQuery, results: queriesResults } in suggestionsSearch"
+          :key="suggestionQuery"
+          :payload="tagging?.searchQueries[suggestionQuery].toolingDisplay ?? emptyTaggingRequest"
+          :event-metadata="{
+            feature: 'ai_carousel',
+            displayOriginalQuery: query || 'ai-carousel-without-query',
+            replaceable: false,
+          }"
+          data-test="ai-carousel-query-display-emitter"
         >
-          <template #sliding-panel-addons="{ arrivedState }">
-            <slot name="sliding-panels-addons" :arrived-state="arrivedState" />
-          </template>
-          <template #sliding-panel-left-button>
-            <slot name="sliding-panels-left-button" />
-          </template>
-          <template #sliding-panel-right-button>
-            <slot name="sliding-panels-right-button" />
-          </template>
-          <div class="x-ai-carousel-suggestion">
-            <ul class="x-ai-carousel-suggestion-results">
-              <DisplayClickProvider
-                v-for="suggestion in suggestionsSearch"
-                :key="suggestion.query"
-                :tooling-display-tagging="
-                  tagging?.searchQueries[suggestion.query].toolingDisplayClick
-                "
-                :tooling-add2-cart-tagging="
-                  tagging?.searchQueries[suggestion.query].toolingDisplayAdd2Cart
-                "
-                result-feature="ai_carousel"
-              >
-                <li
-                  v-for="result in suggestion.results"
-                  :key="result.id"
-                  data-test="ai-carousel-suggestion-result"
+          <div class="x-ai-carousel-suggestion" data-test="ai-carousel-suggestion">
+            <BaseEventButton
+              class="x-ai-carousel-suggestion-query-btn"
+              :events="{ UserAcceptedAQuery: suggestionQuery }"
+            >
+              {{ suggestionQuery }}
+              <ArrowRightIcon class="x-ai-carousel-suggestion-query-btn-icon" />
+            </BaseEventButton>
+
+            <DisplayClickProvider
+              :tooling-display-tagging="tagging?.searchQueries[suggestionQuery].toolingDisplayClick"
+              :tooling-add2-cart-tagging="
+                tagging?.searchQueries[suggestionQuery].toolingDisplayAdd2Cart
+              "
+              result-feature="ai_carousel"
+            >
+              <slot name="sliding-panel" :results="queriesResults">
+                <SlidingPanel
+                  :class="slidingPanelClasses"
+                  :scroll-container-class="slidingPanelContainerClasses"
+                  :button-class="slidingPanelButtonsClasses"
+                  :reset-on-content-change="false"
                 >
-                  <!-- @slot (required) result card -->
-                  <slot name="result" :result="result" />
-                </li>
-              </DisplayClickProvider>
-            </ul>
+                  <template #sliding-panel-addons="{ arrivedState }">
+                    <slot name="sliding-panels-addons" :arrived-state="arrivedState" />
+                  </template>
+                  <template #sliding-panel-left-button>
+                    <slot name="sliding-panels-left-button" />
+                  </template>
+                  <template #sliding-panel-right-button>
+                    <slot name="sliding-panels-right-button" />
+                  </template>
+                  <ul class="x-ai-carousel-suggestion-results">
+                    <li
+                      v-for="result in queriesResults"
+                      :key="result.id"
+                      data-test="ai-carousel-suggestion-result"
+                    >
+                      <!-- @slot (required) result card -->
+                      <slot name="result" :result="result" />
+                    </li>
+                  </ul>
+                </SlidingPanel>
+              </slot>
+            </DisplayClickProvider>
           </div>
-        </SlidingPanel>
-      </slot>
+        </DisplayEmitter>
+        <slot name="suggestions-extra-content" />
+      </div>
+
       <slot name="extra-content" />
       <slot name="cta-button" />
     </div>
@@ -93,48 +129,51 @@ import { useResizeObserver } from '@vueuse/core'
 import { computed, defineComponent, onMounted, ref, watch } from 'vue'
 import {
   AIStarIcon,
+  ArrowRightIcon,
+  BaseEventButton,
   ChangeHeight,
   ChevronDownIcon,
   CollapseHeight,
   DisplayClickProvider,
+  DisplayEmitter,
   SlidingPanel,
 } from '../../../components'
-import DisplayEmitter from '../../../components/display-emitter.vue'
 import { use$x, useState } from '../../../composables'
 import { aiXModule } from '../x-module'
+import AiGroupedCarousel from './ai-grouped-carousel.vue'
 
 export default defineComponent({
   xModule: aiXModule.name,
   components: {
+    AIStarIcon,
+    AiGroupedCarousel,
+    ArrowRightIcon,
+    BaseEventButton,
     ChangeHeight,
+    ChevronDownIcon,
+    CollapseHeight,
     DisplayClickProvider,
     DisplayEmitter,
-    CollapseHeight,
-    AIStarIcon,
-    ChevronDownIcon,
     SlidingPanel,
   },
   props: {
     /* The title text displayed */
-    title: {
-      type: String,
-    },
+    title: String,
     /* The classes added to the sliding panel. */
-    slidingPanelClasses: {
-      type: String,
-    },
+    slidingPanelClasses: String,
     /* The classes added to the sliding panel container. */
-    slidingPanelContainerClasses: {
-      type: String,
-    },
+    slidingPanelContainerClasses: String,
     /* The classes added to the sliding panel buttons. */
-    slidingPanelButtonsClasses: {
-      type: String,
+    slidingPanelButtonsClasses: String,
+    /* Controls whether the carousel should group results or display N carousels per N queries. */
+    group: {
+      type: Boolean,
+      default: true,
     },
   },
   setup(props) {
     const $x = use$x()
-    const { query, isNoResults, suggestionsSearch, queries, tagging } = useState('ai')
+    const { query, suggestionsSearch, queries, tagging } = useState('ai')
     const emptyTaggingRequest: TaggingRequest = { url: '', params: {} }
 
     const titleRef = ref<HTMLElement | null>(null)
@@ -187,7 +226,6 @@ export default defineComponent({
 
     return {
       emptyTaggingRequest,
-      isNoResults,
       isTitleOverflowing,
       queries,
       query,
@@ -201,6 +239,7 @@ export default defineComponent({
   },
 })
 </script>
+
 <style lang="css">
 .x-ai-carousel {
   --color: var(--x-ai-carousel-color, #bbc9cf);
@@ -256,9 +295,35 @@ export default defineComponent({
   transform: rotate(180deg);
 }
 
+.x-ai-carousel-suggestions {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding-bottom: 1rem;
+}
+
 .x-ai-carousel-suggestion {
   display: flex;
+  flex-direction: column;
   gap: 8px;
+}
+
+.x-ai-carousel-suggestion-query-btn {
+  border-color: transparent;
+  background-color: transparent;
+  margin-left: 1rem;
+  margin-right: 1rem;
+  font-weight: 700;
+  width: fit-content;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  cursor: pointer;
+}
+
+.x-ai-carousel-suggestion-query-btn-icon {
+  height: 1rem;
+  aspect-ratio: 1 / 1;
 }
 
 .x-ai-carousel-suggestion-results {