+++ packages/x-components/src/x-modules/ai/components/index.ts
@@ -1,2 +1,1 @@
 export { default as AiCarousel } from './ai-carousel.vue'
-export { default as AiOverview } from './ai-overview.vue'