+++ packages/x-components/src/x-modules/tagging/wiring.ts
@@ -508,10 +508,10 @@ export const taggingWiring = createWiring({
   AiSuggestionsSearchChanged: {
     trackAiSuggestionsSearchWire,
   },
-  UserClickedAnAiOverviewResult: {
+  UserClickedAnAiCarouselResult: {
     trackToolingDisplayClickedWire,
   },
-  UserClickedAnAiOverviewAdd2Cart: {
+  UserClickedAnAiCarouselAdd2Cart: {
     trackToolingAdd2CartWire,
   },
 })