+++ packages/x-adapter-platform/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.0](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.3.0...@empathyco/x-adapter-platform@1.4.0) (2026-05-21)
+
+### Features
+
+* allow empty prices (#2108)
+
+
+## [1.3.0](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.2.1...@empathyco/x-adapter-platform@1.3.0) (2026-05-18)
+
+### Features
+
+* replace X-* Custom headers with requests params for all requests
+
+
+## [1.2.1](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.2.0...@empathyco/x-adapter-platform@1.2.1) (2026-05-07)
+
+**Note:** Version bump only for package @empathyco/x-adapter-platform
+
+
+
+
+
 ## [1.2.0](https://github.com/empathyco/x/compare/@empathyco/x-adapter-platform@1.1.4...@empathyco/x-adapter-platform@1.2.0) (2026-05-05)
 
 ### Features