+++ packages/x-archetype-utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.5](https://github.com/empathyco/x/compare/@empathyco/x-archetype-utils@2.0.4...@empathyco/x-archetype-utils@2.0.5) (2026-05-18)
+
+**Note:** Version bump only for package @empathyco/x-archetype-utils
+
+
+
+
+
+## [2.0.4](https://github.com/empathyco/x/compare/@empathyco/x-archetype-utils@2.0.3...@empathyco/x-archetype-utils@2.0.4) (2026-05-11)
+
+**Note:** Version bump only for package @empathyco/x-archetype-utils
+
+
+
+
+
+## [2.0.3](https://github.com/empathyco/x/compare/@empathyco/x-archetype-utils@2.0.2...@empathyco/x-archetype-utils@2.0.3) (2026-05-07)
+
+**Note:** Version bump only for package @empathyco/x-archetype-utils
+
+
+
+
+
 ## [2.0.2](https://github.com/empathyco/x/compare/@empathyco/x-archetype-utils@2.0.1...@empathyco/x-archetype-utils@2.0.2) (2026-04-20)
 
 **Note:** Version bump only for package @empathyco/x-archetype-utils