+++ packages/x-design-system/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/empathyco/x/compare/@empathyco/x-design-system@1.0.5...@empathyco/x-design-system@1.0.6) (2026-05-18)
+
+**Note:** Version bump only for package @empathyco/x-design-system
+
+
+
+
+
+## [1.0.5](https://github.com/empathyco/x/compare/@empathyco/x-design-system@1.0.4...@empathyco/x-design-system@1.0.5) (2026-05-11)
+
+**Note:** Version bump only for package @empathyco/x-design-system
+
+
+
+
+
+## [1.0.4](https://github.com/empathyco/x/compare/@empathyco/x-design-system@1.0.3...@empathyco/x-design-system@1.0.4) (2026-05-07)
+
+**Note:** Version bump only for package @empathyco/x-design-system
+
+
+
+
+
 ## [1.0.3](https://github.com/empathyco/x/compare/@empathyco/x-design-system@1.0.2...@empathyco/x-design-system@1.0.3) (2026-05-07)
 
 ### Bug Fixes